training/requirements.txt
joblib==1.3.1
kafka-python==2.0.2
numpy==1.25.1
pandas==2.0.3
pyRAPL==0.2.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3